# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Private RAG Apps API",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health Check",
        "description": "APIの死活監視のためのヘルスチェックエンドポイント",
        "operationId": "health_check_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/conversations": {
      "get": {
        "summary": "List Conversations",
        "description": "保存されているすべての会話履歴の一覧を更新日時順に取得する",
        "operationId": "list_conversations_api_conversations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationResponse"
                  },
                  "type": "array",
                  "title": "Response List Conversations Api Conversations Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Conversation",
        "description": "新規の会話（チャットセッション）を作成し、IDと初期タイトルを返す",
        "operationId": "create_conversation_api_conversations_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/conversations/{conversation_id}": {
      "get": {
        "summary": "Get Conversation",
        "description": "指定されたIDの会話詳細と、それに紐づくメッセージ履歴を取得する",
        "operationId": "get_conversation_api_conversations__conversation_id__get",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "summary": "Chat",
        "description": "ユーザーからのメッセージを受け取り、会話履歴を踏まえた検索と回答生成を行う。\n生成結果はServer-Sent Events (SSE)を用いてストリーミングで返却し、終了時にメッセージをDBへ一括保存する。",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "ユーザーからの質問メッセージ"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "継続する会話のID。未指定の場合は新規会話を作成する"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ConversationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "会話ID"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "会話タイトル（初回メッセージの先頭から自動生成）"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "最終更新日時"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "updated_at"
        ],
        "title": "ConversationResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}